1. Add support to arpack-ng. Users can choose to apply Arnoldi using arpack-ng. 2. Change the METIS interface. Now we use INT64 + FLOAT as default option. We use float instaed of double when calling METIS functions.    Users can enable INT32 by compiling with -DPARGEMSLR_INT32. Currently the hypre interface only works with INT64. 3. Add several matrix operation functions. 4. Add several MPI interface functions.

+++ ParGeMSLR/SRC/matrices/coo_matrix.cpp
@@ -270,7 +270,7 @@ namespace pargemslr
    template <typename T>
    long int CooMatrixClass<T>::GetNumNonzeros() const
    {
-      return this->_nnz;
+      return (long int)this->_nnz;
    }
    template long int CooMatrixClass<float>::GetNumNonzeros() const;
    template long int CooMatrixClass<double>::GetNumNonzeros() const;
